# src/main/java/br/com/cartpurchase/model/dto/EmailDTO.java
package br.com.cartpurchase.model.dto;

import lombok.Data;

import javax.validation.constraints.Email;
import javax.validation.constraints.NotBlank;
import java.io.Serializable;

@Data
public class EmailDTO implements Serializable {

    @NotBlank
    private String ownerRef;
    @NotBlank
    @Email
    private String emailFrom;
    @NotBlank
    @Email
    private String emailTo;
    private String subject;
    private String text;

}